* Added ObjManager::list_add_obj() Adds an Obj into a U6LList combining   with existing objects if stackable. * Added ActorManager::resurrect_actor() This converts a dead body Obj   back into the corresponding Actor record. It also transfers the dead   body's contents back into the Actor inventory. * Added obj_stackable[] table into ObjManager. This is now used by   ObjManager::is_stackable() to determine if an Obj is stackable. * Changed Obj.qty from uint8 to uint16. * Added support for Obj quantities greater than 255. * Added FM-Towns audio dialog support. To use this feature put the   following line inside your <ultima6></ultima6> config block.
  <townsdir>/the/path/to/your/fmtowns/u6/dir</townsdir>

  The FM-Towns audio is sampled at 16KHz and will need up converting to
  play correctly with the Nuvie audio system. This hasn't been done yet
  so the voices will sound wrong. If you want to hear how the voices
  should sound. Change the audio_rate in SoundManager::initAudio() from
  44100 to 16000.
* Added ConverseSpeech class. This is used for playing the FM-Towns
  audio samples when in Talk mode.
* Hooked up the actor ressurection action in the Converse system.
* Added U6LList::findPos() This method finds the position in the list
  holding the reference pointed to by void *data.
* Added InventoryWidget::display_special_char() This prints the special
  char under keys in U6.
* Added the inventory key letter fonts to InventoryFont.h

+++ Event.h
@@ -148,10 +148,10 @@ class Event : public GUI_CallBack
  void party_mode();
  bool ready(Obj *obj);
  bool unready(Obj *obj);
- bool drop_select(Obj *obj, uint8 qty = 0);
- bool drop_count(uint8 qty);
+ bool drop_select(Obj *obj, uint16 qty = 0);
+ bool drop_count(uint16 qty);
  bool drop();
- bool drop(Obj *obj, uint8 qty, uint16 x, uint16 y);
+ bool drop(Obj *obj, uint16 qty, uint16 x, uint16 y);
  bool drop(uint16 x, uint16 y) { return(drop(use_obj, drop_qty, x, y)); }
  void walk_to_mouse_cursor(uint32 mx, uint32 my);
  void multiuse(uint16 wx, uint16 wy);
